+++ apiml-common/build.gradle
@@ -8,6 +8,9 @@ dependencies {
     implementation libs.spring.boot.starter.actuator
     implementation libs.spring.boot.starter.web
     implementation libs.spring.cloud.starter.eureka.client
+    implementation libs.spring.cloud.gateway.server
+    implementation libs.spring.webflux
+    implementation libs.netty.reactor.http
     implementation libs.eureka.core
 
     implementation libs.apache.commons.lang3
