+++ gateway-service/src/main/java/org/zowe/apiml/gateway/config/ConnectionsConfig.java
@@ -22,7 +22,6 @@ import io.github.resilience4j.circuitbreaker.CircuitBreakerConfig;
 import io.github.resilience4j.timelimiter.TimeLimiterConfig;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslContextBuilder;
-import io.netty.resolver.DefaultAddressResolverGroup;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.Delegate;
 import lombok.extern.slf4j.Slf4j;
@@ -35,15 +34,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.web.ServerProperties;
 import org.springframework.cloud.circuitbreaker.resilience4j.ReactiveResilience4JCircuitBreakerFactory;
 import org.springframework.cloud.circuitbreaker.resilience4j.Resilience4JConfigBuilder;
 import org.springframework.cloud.client.circuitbreaker.Customizer;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
-import org.springframework.cloud.gateway.config.HttpClientCustomizer;
-import org.springframework.cloud.gateway.config.HttpClientFactory;
+
 import org.springframework.cloud.gateway.config.HttpClientProperties;
-import org.springframework.cloud.gateway.config.HttpClientSslConfigurer;
 import org.springframework.cloud.gateway.filter.headers.HttpHeadersFilter;
 import org.springframework.cloud.netflix.eureka.CloudEurekaClient;
 import org.springframework.cloud.netflix.eureka.EurekaClientConfigBean;
@@ -57,12 +53,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.DependsOn;
-import org.springframework.context.annotation.Primary;
-import org.springframework.http.client.reactive.ReactorClientHttpConnector;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.cors.reactive.CorsWebFilter;
-import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.server.WebFilter;
 import org.springframework.web.util.UriComponentsBuilder;
 import org.zowe.apiml.config.AdditionalRegistration;
@@ -170,45 +163,6 @@ public class ConnectionsConfig {
     @VisibleForTesting
     X509KeyManager x509KeyManagerSelectedAlias(KeyManagerFactory keyManagerFactory) {
         return new X509KeyManagerSelectedAlias(keyManagerFactory, config.getKeyAlias());
-    }
-
-    /**
-     * @return io.netty.handler.ssl.SslContext for http client.
-     */
-    SslContext getSslContext(boolean setKeystore) {
-        try {
-            SslContextBuilder builder = SslContextBuilder.forClient();
-
-            KeyStore trustStore = SecurityUtils.loadKeyStore(
-                config.getTrustStoreType(), config.getTrustStorePath(), config.getTrustStorePassword());
-            TrustManagerFactory trustManagerFactory = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-            trustManagerFactory.init(trustStore);
-            builder.trustManager(trustManagerFactory);
-
-            KeyManagerFactory keyManagerFactory = KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
-            if (setKeystore) {
-                log.info("Loading keystore: {}: {}", config.getKeyStoreType(), config.getKeyStorePath());
-                KeyStore keyStore = SecurityUtils.loadKeyStore(
-                    config.getKeyStoreType(), config.getKeyStorePath(), config.getKeyStorePassword());
-                keyManagerFactory.init(keyStore, config.getKeyStorePassword());
-                builder.keyManager(x509KeyManagerSelectedAlias(keyManagerFactory));
-            } else {
-                KeyStore emptyKeystore = KeyStore.getInstance(KeyStore.getDefaultType());
-                emptyKeystore.load(null, null);
-                keyManagerFactory.init(emptyKeystore, null);
-                builder.keyManager(keyManagerFactory);
-            }
-
-            if (config.isVerifySslCertificatesOfServices() && config.isNonStrictVerifySslCertificatesOfServices()) {
-                builder.endpointIdentificationAlgorithm(null);
-            }
-
-            return builder.build();
-        } catch (Exception e) {
-            apimlLog.log("org.zowe.apiml.common.sslContextInitializationError", e.getMessage());
-            throw new HttpsConfigError("Error initializing SSL Context: " + e.getMessage(), e,
-                HttpsConfigError.ErrorCode.HTTP_CLIENT_INITIALIZATION_FAILED, config.httpsConfig());
-        }
     }
 
     @Bean(destroyMethod = "shutdown", name = "eurekaClient")
@@ -337,38 +291,6 @@ public class ConnectionsConfig {
                 TimeLimiterConfig.custom()
                     .timeoutDuration(Duration.ofMillis(config.getRequestConnectionTimeout()))
                 .build()).build());
-    }
-
-    @Bean
-    HttpClientFactory gatewayHttpClientFactory(
-        HttpClientProperties properties,
-        ServerProperties serverProperties, List<HttpClientCustomizer> customizers,
-        HttpClientSslConfigurer sslConfigurer
-    ) {
-        SslContext sslContext = getSslContext(false);
-        return new HttpClientFactory(properties, serverProperties, sslConfigurer, customizers) {
-            @Override
-            protected HttpClient createInstance() {
-                return super.createInstance()
-                    .secure(sslContextSpec -> sslContextSpec.sslContext(sslContext))
-                    .resolver(DefaultAddressResolverGroup.INSTANCE);
-            }
-        };
-    }
-
-    @Bean
-    @Primary
-    WebClient webClient(HttpClient httpClient) {
-        return WebClient.builder()
-            .clientConnector(new ReactorClientHttpConnector(getHttpClient(httpClient, false)))
-            .build();
-    }
-
-    @Bean
-    WebClient webClientClientCert(HttpClient httpClient) {
-        return WebClient.builder()
-            .clientConnector(new ReactorClientHttpConnector(getHttpClient(httpClient, true)))
-            .build();
     }
 
     @Bean
